Separate call control logic from the dialler
Group components into separate modules within a package. This encourages
separation of campaign logic/configuration from per call processing.
